Conflicts:
	Sources/Internal/Render/2D/FTFont.cpp
	Sources/Internal/Render/2D/TextBlock.cpp

Former-commit-id: 0b7f6fa87000a72be3ecac762ab46a450bf0a5fa

+++ Sources/Internal/Render/RHI/DX9/rhi_QueryBufferDX9.cpp
@@ -31,7 +31,7 @@
     #include "rhi_DX9.h"
 
     #include "Debug/DVAssert.h"
-    #include "FileSystem/Logger.h"
+    #include "Logger/Logger.h"
 using DAVA::Logger;
 
     #include "_dx9.h"
@@ -96,7 +96,7 @@ dx9_QueryBuffer_Delete(Handle handle)
             cmd.push_back(c);
         }
 
-        ExecDX9(&cmd[0], cmd.size());
+        ExecDX9(&cmd[0], static_cast<uint32>(cmd.size()));
         buf->query.clear();
     }
